+++ app/Listings/add.listing.controller.js
@@ -1,35 +1,36 @@
-(function() {
+(function () {
     'use strict';
 
-   angular
+    angular
         .module('app')
         .controller('AddListingsController', AddListingsController);
 
-   AddListingsController.$inject = ['ListingFactory'];
-    
-   function AddListingsController(ListingFactory) {
+    AddListingsController.$inject = ['ListingFactory'];
+
+    function AddListingsController(ListingFactory) {
         var AddListCtrl = this;
-        
-       AddListCtrl.listObject = {};
+
+        AddListCtrl.listObject = {};
         AddListCtrl.listObject.description = "";
         AddListCtrl.listObject.price = 0;
         AddListCtrl.listObject.address = "";
         AddListCtrl.listObject.city = "";
         AddListCtrl.listObject.state = "";
-        AddListCtrl.listObject.zipcode = 0;
+        AddListCtrl.listObject.zipCode = 0;
         AddListCtrl.listObject.picture = '/imagepath';
+        AddListCtrl.listObject.userID = 1;
 
-       ////////////////
+        ////////////////
 
-       vm.addListing = function (listingInfo) {
+        AddListCtrl.addListing = function (listingInfo) {
             ListingFactory
-            .postListing (listingInfo)
-            .then (function (response) {
-                console.log(response);
+                .postListing(listingInfo)
+                .then(function (response) {
+                    console.log(response);
 
-           }, function (error) {
-                console.log(error);
-            })
+                }, function (error) {
+                    console.log(error);
+                })
         }
     }
 })();
