+++ app/Listings/add.listings.html
@@ -6,36 +6,36 @@
     <form class="form-horizontal">
         <div class="form-group">
             <label rows="4" cols="50" for="propertyName" class="control-label col-lg-2"> Description </label>
-            <textarea ng-model="vm.propObject.description" type="text" required></textarea>
+            <textarea ng-model="AddListCtrl.listObject.description" type="text" required></textarea>
         </div>
         <div class="form-group">
             <label for="propertyName" class="control-label col-lg-2">Rent</label>
-            <input ng-model="vm.propObject.price" type="text" required>
+            <input ng-model="AddListCtrl.listObject.price" type="text" required>
         </div>
         <div class="form-group">
             <label for="propertyName" class="control-label col-lg-2"> Address </label>
-            <input ng-model="vm.propObject.address" type="text" required>
+            <input ng-model="AddListCtrl.listObject.address" type="text" required>
         </div>
         <div class="form-group">
             <label for="propertyName" class="control-label col-lg-2"> City </label>
-            <input ng-model="vm.propObject.city" type="text" required>
+            <input ng-model="AddListCtrl.listObject.city" type="text" required>
         </div>
         <div class="form-group">
             <label for="propertyName" class="control-label col-lg-2"> State </label>
-            <input ng-model="vm.propObject.state" type="text" required>
+            <input ng-model="AddListCtrl.listObject.state" type="text" required>
         </div>
-        <div class=“form-group”>
-            <label for=“propertyName” class=“control-label col-lg-2”> Zip Code </label>
-            <input ng-model=“vm.propObject.zipCode” type="text" required>
+        <div class="form-group">
+            <label for="propertyName" class="control-label col-lg-2"> Zip Code </label>
+            <input ng-model="AddListCtrl.listObject.zipCode" type="text" required>
         </div>
         <div class="form-group">
             <label for="exampleInputFile" class="control-label col-lg-2">File input</label>
-            <input type="file" ng-model="vm.propObject.picture" class="form-control-file" id="exampleInputFile" aria-describedby="fileHelp">
+            <input type="file" ng-model="AddListCtrl.listObject.picture" class="form-control-file" id="exampleInputFile" aria-describedby="fileHelp">
             <small id="fileHelp" class="form-text text-muted">Add a photo here to display with your listing.</small>
         </div>
         <div class="form-group">
             <div class="col-sm-offset-2 col-sm-10">
-                <button type="submit" ng-click="vm.addListing(vm.listObject)">Add Your Listing</button>
+                <button type="submit" ng-click="AddListCtrl.addListing(AddListCtrl.listObject)">Add Your Listing</button>
             </div>
     </form>
