+++ index.html
@@ -16,12 +16,12 @@
 <body>
 
   <div id="topbar">
-    <h1 id="MainTitle"> 93' Til Brewing</h1>
+    <h1 id="MainTitle">Welcome to RoomMe</h1>
     <div class="topnav" id="myTopnav">
-      <img src="houseoutline.jpg" style="width:50px;height:50px;"/>
       <a ui-sref="profile">Profile</a>
       <a ui-sref="search">Listings</a>
       <a ui-sref="register">Log In</a>
+      <a ui-sref="addListing">Add Listing</a>
     </div>
   </div>
   
@@ -51,12 +51,12 @@
 <script src="/app/Factories/owner.listing.factory.js"></script>
 <script src="/app/Factories/search.factory.js"></script>
 <script src="/app/Factories/user.factory.js"></script>
+<script src="/app/Profile/profile.controller.js"></script>
 <script src="/app/Listings/add.listing.controller.js"></script>
 <script src="/app/Listings/owner.controller.js"></script>
-<script src="/app/Messages/message.controller.js"></script>
-<script src="/app/Profile/profile.controller.js"></script>
 <script src="/app/Search/search.controller.js"></script>
 <script src="/app/Users/signin.controller.js"></script>
+<script src="/app/Messages/message.controller.js"></script>
 <!-- endinject -->
 </body>
 </html>
